♻️ refactor: 리뷰 반영하여 리팩터링
