+++ apps/web/src/app/(needLogin)/mypage/setting/service/_components/index.ts
@@ -1,3 +1,2 @@
-export * from './ServiceValidator/ServiceValidator';
-export * from './ServiceValidator/ServiceValidator.types';
+export * from './ServiceForm/ServiceForm';
 export * from './UserInquiryHistory/UserInquiryHistory';
